move folder view into fs/folder
