+++ shared/fs/common/open-hoc.js
@@ -4,7 +4,6 @@ import * as I from 'immutable'
 import * as FsGen from '../../actions/fs-gen'
 import * as Constants from '../../constants/fs'
 import * as Types from '../../constants/types/fs'
-import {memoize} from '../../util/memoize'
 import {namedConnect} from '../../util/container'
 
 type OwnProps = {
@@ -34,10 +33,8 @@ const isFolder = (stateProps, ownProps: OwnProps) =>
   Types.getPathLevel(ownProps.path) <= 3 ||
   stateProps._pathItems.get(ownProps.path, Constants.unknownPathItem).type === 'folder'
 
-const canOpenInDestinationPicker = memoize(
-  (stateProps, ownProps) =>
-    !isFolder(stateProps, ownProps) || stateProps._moveOrCopy.sourceItemPath === ownProps.path
-)
+const canOpenInDestinationPicker = (stateProps, ownProps) =>
+  isFolder(stateProps, ownProps) && stateProps._moveOrCopy.sourceItemPath !== ownProps.path
 
 type MergedProps = OwnProps & {
   onOpen: ?() => void,
@@ -47,8 +44,8 @@ const mergeProps = (stateProps, dispatchProps, ownProps: OwnProps): MergedProps 
   onOpen:
     typeof ownProps.destinationPickerIndex === 'number'
       ? canOpenInDestinationPicker(stateProps, ownProps)
-        ? null
-        : dispatchProps._destinationPickerGoTo
+        ? dispatchProps._destinationPickerGoTo
+        : null
       : dispatchProps._open,
   // We need the inexact spread here because this is a HOC. As such, it must
   // pass down any OwnProps to composed components, even if the HOC typing
